flask/requirements.txt
alembic==1.4.3
Authlib==0.15.2
cachetools==4.1.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
google-api-core==1.23.0
google-auth==1.24.0
google-cloud-vision==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.33.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
libcst==0.3.14
Mako==1.1.3
MarkupSafe==1.1.1
mypy-extensions==0.4.3
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
proto-plus==1.11.0
protobuf==3.14.0
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.20
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
uWSGI==2.0.19.1
Werkzeug==1.0.1
